{input,mixer}/alsa: move code to lib/alsa/NonBlock.cxx
